refactor(ingest): split lineage URN casing processor into a package
AutoResolveLineageUrnsProcessor mixed two concerns in one 712-line module:
what to rewrite (which aspects carry upstream dataset references, how to
rewrite them, how to count and report the outcome) and how to find the URN
DataHub actually stores (download each configured platform's catalog, then
match three casing candidates locally).

Only the second is about to change, so separate them behind a Protocol:

  models.py     MatchType, Resolution, ResolutionStrategy, the report
  processor.py  everything independent of how a reference resolves
  bulk.py       today's catalog download and casing match

The processor now talks to a ResolutionStrategy in five places and names a
concrete strategy in exactly one, so an alternative resolution path plugs in
without touching the rewriters, and removing one later is a file deletion
plus a single line.

No behaviour change. The package keeps the old module's import path, so no
call site moves. One relocation worth noting: the unresolved-refs report
sample is now appended by the processor's counting paths rather than inside
the strategy, because create() assigns self.report only after __init__
returns, so a strategy built in __init__ cannot receive it. Every path that
returned UNRESOLVED already went through one of those counters and bulk
returns the reference URN unchanged on a miss, so the sample is identical;
a new assertion pins it, since that line had no coverage.

bulk.py is imported lazily, keeping sqlglot off the module-load path that
every source hits (test_module_import_does_not_pull_sqlglot).
