+++ metadata-ingestion/src/datahub/ingestion/workunit_processors/auto_resolve_lineage_urns/__init__.py
@@ -1,0 +1,11 @@
+from datahub.ingestion.workunit_processors.auto_resolve_lineage_urns.models import (
+    AutoResolveLineageUrnsProcessorReport,
+)
+from datahub.ingestion.workunit_processors.auto_resolve_lineage_urns.processor import (
+    AutoResolveLineageUrnsProcessor,
+)
+
+__all__ = [
+    "AutoResolveLineageUrnsProcessor",
+    "AutoResolveLineageUrnsProcessorReport",
+]
