+++ metadata-ingestion/src/datahub/ingestion/workunit_processors/auto_resolve_lineage_urns/processor.py
@@ -1,22 +1,17 @@
-# NOTE: `from __future__ import annotations` keeps the schema_resolver type hints
-# (imported only under TYPE_CHECKING) as strings, so importing this module does not
-# pull in sqlglot. This module is imported eagerly on every source's
-# get_workunit_processors() path, so module load must stay sqlglot-free (guarded by
-# test_module_import_does_not_pull_sqlglot). The sqlglot-heavy schema_resolver imports
-# are therefore deferred to a single chokepoint in __init__, which runs only after
-# should_enable() confirms the feature is on and a graph exists — off the module-load
-# path, but honest about the dependency (see __init__).
+# NOTE: `from __future__ import annotations` keeps the schema_resolver type hints (imported
+# only under TYPE_CHECKING) as strings, so importing this module does not pull in sqlglot.
+# This module is imported eagerly on every source's get_workunit_processors() path, so
+# module load must stay sqlglot-free (guarded by test_module_import_does_not_pull_sqlglot).
+# The sqlglot-heavy imports are therefore deferred: match_columns_to_schema to a chokepoint
+# in __init__, and the strategy module to the same place — both run only after
+# should_enable() confirms the feature is on and a graph exists.
 from __future__ import annotations
 
-import logging
-from dataclasses import dataclass, field
 from typing import (
     TYPE_CHECKING,
     Callable,
-    Dict,
     Iterable,
     List,
-    Literal,
     Optional,
     Set,
     Tuple,
@@ -35,7 +30,15 @@ from datahub.ingestion.api.workunit import MetadataWorkUnit
 from datahub.ingestion.api.workunit_processor import (
     WorkunitProcessor,
     WorkunitProcessorContext,
-    WorkunitProcessorReport,
+)
+from datahub.ingestion.workunit_processors.auto_resolve_lineage_urns.models import (
+    EXACT,
+    NORMALIZED,
+    UNRESOLVED,
+    AutoResolveLineageUrnsProcessorReport,
+    MatchType,
+    Resolution,
+    ResolutionStrategy,
 )
 from datahub.metadata.schema_classes import (
     ChangeTypeClass,
@@ -44,69 +47,15 @@ from datahub.metadata.schema_classes import (
     DataJobInputOutputClass,
     EdgeClass,
     FineGrainedLineageClass,
-    LineageMatchTypeClass,
     MetadataChangeProposalClass,
     UpstreamLineageClass,
     _Aspect,
 )
-from datahub.metadata.urns import DataPlatformUrn, DatasetUrn, SchemaFieldUrn
-from datahub.utilities.lossy_collections import LossyList
+from datahub.metadata.urns import DatasetUrn, SchemaFieldUrn
 from datahub.utilities.urns.error import InvalidUrnError
 
 if TYPE_CHECKING:
-    from datahub.ingestion.graph.client import DataHubGraph
-    from datahub.ingestion.run.pipeline_config import UpstreamPlatformCasing
-    from datahub.sql_parsing.schema_resolver import SchemaInfo, SchemaResolver
-
-logger = logging.getLogger(__name__)
-
-# Above this many URNs per platform, the bulk-loaded SchemaResolver cache is large
-# enough to warrant an explicit heads-up to operators rather than letting it surface as
-# unexplained memory pressure. (A disk-backed, casing-aware resolver owned by
-# SchemaResolver is the planned follow-up; see the normalizedUrn backlog task.)
-_CATALOG_SIZE_WARN_THRESHOLD = 500_000
-
-# The closed set of matchType verdicts, as a Literal so the if/elif verdict chains that
-# drive correctness can be typo- and exhaustiveness-checked. LineageMatchTypeClass
-# renders these as plain ``str`` (codegen), so we bind Literal-typed aliases and assert
-# they stay in sync with the generated class.
-MatchType = Literal["EXACT", "NORMALIZED", "UNRESOLVED"]
-_EXACT: MatchType = "EXACT"
-_NORMALIZED: MatchType = "NORMALIZED"
-_UNRESOLVED: MatchType = "UNRESOLVED"
-assert (_EXACT, _NORMALIZED, _UNRESOLVED) == (
-    LineageMatchTypeClass.EXACT,
-    LineageMatchTypeClass.NORMALIZED,
-    LineageMatchTypeClass.UNRESOLVED,
-), "MatchType literals drifted from LineageMatchTypeClass"
-
-
-@dataclass
-class AutoResolveLineageUrnsProcessorReport(WorkunitProcessorReport):
-    """Report for AutoResolveLineageUrnsProcessor metrics."""
-
-    num_dataset_urns_normalized: int = 0  # Upstream dataset URNs rewritten
-    num_column_urns_normalized: int = 0  # Fine-grained field URNs rewritten
-    num_refs_unchanged: int = 0  # Left as-is (exact match, or out of scope)
-    num_refs_unresolved: int = 0  # Configured platform, no unique match (flagged)
-    num_exceptions: int = 0  # Failed to process a workunit
-    # Lineage aspect emitted as a PATCH (not UPSERT); can't be reconciled, so skipped.
-    num_patch_lineage_skipped: int = 0
-    num_workunits_with_lineage_aspect: int = 0
-    num_workunits_modified: int = 0
-    # Bounded sample of references left UNRESOLVED, alongside the num_refs_unresolved
-    # count, so the report shows *which* lineage looks broken, not just how much.
-    unresolved_refs_sample: LossyList[str] = field(default_factory=LossyList)
-
-
-@dataclass
-class _Resolution:
-    """Outcome of resolving one dataset URN against the configured platform(s)."""
-
-    urn: str  # The (possibly rewritten) URN to emit.
-    schema: Optional[SchemaInfo]  # Cached schema of the resolved entity, if known.
-    # EXACT / NORMALIZED / UNRESOLVED / None (no reconciliation performed).
-    match_type: Optional[MatchType]
+    from datahub.sql_parsing.schema_resolver import SchemaInfo
 
 
 def _parent_dataset_urn(field_urn: str) -> Optional[str]:
@@ -160,60 +109,35 @@ class AutoResolveLineageUrnsProcessor(
 
     Heals casing mismatches between sources (e.g. a lowercase-stored Snowflake table
     referenced in a different casing by a BI tool) that would otherwise create two
-    disconnected lineage nodes. For each configured upstream platform it bulk-loads that
-    platform's URNs and schemas once (via ``SchemaResolverProvider``) and resolves every
-    reference locally via ``SchemaResolver.resolve_table`` (which tries the original,
-    lowercased, and mixed-instance casings — see ``_resolve_dataset``) against the casing
-    DataHub already stores — table-level (``UpstreamLineage``,
-    ``DashboardInfo``) and column-level (``FineGrainedLineage`` field paths). Broader
-    any-casing resolution is a tracked SchemaResolver follow-up (the ``normalizedUrn``
-    aspect).
+    disconnected lineage nodes. Reconciliation covers table-level (``UpstreamLineage``,
+    ``DashboardInfo``, ``ChartInfo``, ``DataJobInputOutput``) and column-level
+    (``FineGrainedLineage`` field paths) references.
 
-    Only references *to* warehouse assets found in this source's metadata are fixed;
-    the entity the aspect is attached to and downstream fields are never touched. It
-    must be enabled on BI-tool / cross-platform ingestions and configured with the
-    upstream platform(s) — never on the warehouse ingestion, whose reported casing and
-    identity must be respected.
+    Finding the URN DataHub actually stores is delegated to a :class:`ResolutionStrategy`;
+    this class owns everything that is the same regardless of how a reference resolves —
+    which aspects carry upstream references, how they are rewritten, and how verdicts are
+    counted and reported.
+
+    Only references *to* warehouse assets are fixed; the entity the aspect is attached to
+    and downstream fields are never touched. It must be enabled on BI-tool / cross-platform
+    ingestions — never on the warehouse ingestion, whose reported casing and identity must
+    be respected.
     """
 
     def __init__(self, ctx: WorkunitProcessorContext) -> None:
         super().__init__(ctx)
-        graph = ctx.pipeline_context.graph
-        # assert for the type checker.
-        assert graph is not None
-        self._graph: DataHubGraph = graph
-        self._config: List[UpstreamPlatformCasing] = (
-            ctx.pipeline_context.flags.auto_resolve_lineage_urns.upstream_platforms
+        # Deferred imports, for the same reason the module docstring gives: both of these
+        # reach sqlglot-backed code, and this __init__ runs only once should_enable() has
+        # confirmed the feature is on.
+        from datahub.ingestion.workunit_processors.auto_resolve_lineage_urns.bulk import (
+            BulkCatalogStrategy,
         )
-        # Resolve the sqlglot-backed schema_resolver callables once, here — a single
-        # honest chokepoint rather than imports buried in two leaf methods. Deferred into
-        # __init__ (not module level) so importing this module stays sqlglot-free
-        # (guarded by test_module_import_does_not_pull_sqlglot); __init__ runs only after
-        # should_enable() confirmed the feature is on. The sqlglot dependency itself is
-        # validated up front by AutoResolveLineageUrnsConfig (fail-fast at config parse
-        # when enabled), so these imports are guaranteed to succeed here.
         from datahub.sql_parsing.schema_resolver import match_columns_to_schema
-        from datahub.sql_parsing.schema_resolver_provider import provide_schema_resolver
 
-        self._provide_schema_resolver: Callable[..., SchemaResolver] = (
-            provide_schema_resolver
-        )
         self._match_columns_to_schema: Callable[[SchemaInfo, List[str]], List[str]] = (
             match_columns_to_schema
         )
-        # Per-platform SchemaResolvers, bulk-initialized up front by _load_catalogs().
-        # Casing matching is delegated to SchemaResolver.resolve_table (which tries the
-        # reference's original, lowercased, and mixed-instance casings — see
-        # _resolve_dataset) — the processor keeps no parallel casing index of its own.
-        # Broader casing coverage (a non-lowercase table stored as Pascal / Mixed /
-        # arbitrary) and casing-aware resolution are a tracked SchemaResolver follow-up
-        # (the planned `normalizedUrn` aspect). Until that lands, only casings resolve_table
-        # covers are reconciled here.
-        self._resolvers_by_platform: Dict[str, List["SchemaResolver"]] = {}
-        # Platforms actually referenced by this source's lineage, so
-        # _warn_unmatched_platforms can flag configured platforms that no reference used
-        # (usually a case/spelling typo in the config).
-        self._seen_reference_platforms: Set[str] = set()
+        self._strategy: ResolutionStrategy = BulkCatalogStrategy(ctx)
         # (aspect class -> in-place normalizer, returns True iff it mutated the aspect).
         # These are the aspects a BI / orchestration source emits that carry *upstream
         # dataset* references — the only refs affected by cross-source casing mismatch:
@@ -224,7 +148,7 @@ class AutoResolveLineageUrnsProcessor(
         # check for MCE/MCPW (live aspect) and, for a raw MCP, short-circuits on aspectName
         # before any deserialization — so a work unit is deserialized at most once (for the
         # aspect it actually carries), and covering four vs. one adds only three constant
-        # comparisons. The one real cost, the up-front catalog load, is independent of this.
+        # comparisons.
         # Callable[..., bool] (not Callable[[_Aspect], bool]): each normalizer takes a
         # specific aspect subtype, and function args are contravariant, so the precise
         # signature won't accept them in a heterogeneous table (mypy list-item error).
@@ -239,7 +163,6 @@ class AutoResolveLineageUrnsProcessor(
         self._lineage_aspect_names: Set[str] = {
             aspect_cls.ASPECT_NAME for aspect_cls, _ in self._normalizers
         }
-        self._load_catalogs()
 
     @classmethod
     def should_enable(cls, ctx: WorkunitProcessorContext) -> bool:
@@ -277,7 +200,7 @@ class AutoResolveLineageUrnsProcessor(
                     exc=e,
                 )
             yield wu
-        self._warn_unmatched_platforms()
+        self._strategy.finish()
         self._warn_unresolved_refs()
         self._warn_patch_lineage_skipped()
 
@@ -316,34 +239,6 @@ class AutoResolveLineageUrnsProcessor(
             if normalize(aspect):
                 self._write_back_if_mcp(wu, aspect)
                 self.report.num_workunits_modified += 1
-
-    def _warn_unmatched_platforms(self) -> None:
-        """Surface configured platforms that no reference used, in the pipeline report.
-
-        The usual cause is a case/spelling mismatch in the config (platform names are
-        compared case-sensitively). Emitting it as a structured (UI-visible) warning lets
-        the operator either fix the platform name or drop the platform if it isn't
-        actually referenced by this source.
-        """
-        # Defense in depth: this runs outside the per-workunit try/except, so guard
-        # against a non-list config (should_enable already fails closed on a mock ctx).
-        if not isinstance(self._config, list):
-            return
-        unmatched = {
-            entry.platform for entry in self._config
-        } - self._seen_reference_platforms
-        if not unmatched:
-            return
-        self.ctx.source_report.warning(
-            title="Configured upstream platform matched no lineage references",
-            message="An upstream platform configured under auto_resolve_lineage_urns was "
-            "not referenced by any lineage in this run, so nothing was reconciled for it. "
-            "Platform names are matched case-sensitively against the dataset URN's "
-            "platform (e.g. 'snowflake', not 'Snowflake') — fix the name if it's a typo, "
-            "or remove the platform from upstream_platforms if this source doesn't "
-            "reference it.",
-            context=f"{sorted(unmatched)}",
-        )
 
     def _warn_unresolved_refs(self) -> None:
         """Surface UNRESOLVED references in the pipeline report, once, aggregated.
@@ -394,160 +289,22 @@ class AutoResolveLineageUrnsProcessor(
         if isinstance(wu.metadata, MetadataChangeProposalClass):
             wu.metadata.aspect = _make_generic_aspect(aspect)
 
-    # --- resolution -------------------------------------------------------------
-
-    def _load_catalogs(self) -> None:
-        """Bulk-load every configured platform's SchemaResolver once, up front.
-
-        ``provide_schema_resolver`` does a single bulk scroll per (platform, instance, env)
-        and is globally cached, warming each resolver's schema cache so the per-reference
-        ``resolve_table`` calls in _resolve_dataset stay local (no per-reference round
-        trips). Matching itself is delegated to SchemaResolver; this processor keeps no
-        casing index of its own.
-        """
-        entries_by_platform: Dict[str, List["UpstreamPlatformCasing"]] = {}
-        for entry in self._config:
-            entries_by_platform.setdefault(entry.platform, []).append(entry)
-
-        for platform, entries in entries_by_platform.items():
-            # Emitted before the (potentially long, paginated) fetch so operators see a
-            # signal during the stall, not only after.
-            logger.info(
-                f"Loading '{platform}' catalog from DataHub for lineage casing "
-                f"reconciliation; this may take a while on large warehouses..."
-            )
-            resolvers: List[SchemaResolver] = []
-            try:
-                for entry in entries:
-                    resolvers.append(
-                        self._provide_schema_resolver(
-                            graph=self._graph,
-                            platform=entry.platform,
-                            platform_instance=entry.platform_instance,
-                            env=entry.env,
-                        )
-                    )
-            except Exception as e:
-                # A catalog-load failure must not crash the pipeline: report it and leave
-                # the platform unloaded, so its references are emitted unchanged.
-                self.ctx.source_report.warning(
-                    title="Lineage URN casing: upstream catalog not loaded",
-                    message="Failed to bulk-load an upstream platform's catalog from "
-                    "DataHub; references to it are emitted unchanged.",
-                    context=platform,
-                    exc=e,
-                )
-                continue
-            # The resolver caches are held for the pipeline's lifetime; log their size,
-            # escalating to WARNING once large enough to matter.
-            count = sum(len(r.get_urns()) for r in resolvers)
-            message = (
-                f"Loaded {count} '{platform}' dataset URNs for lineage casing "
-                f"reconciliation."
-            )
-            if count > _CATALOG_SIZE_WARN_THRESHOLD:
-                logger.warning(
-                    f"{message} This is a large catalog and may use significant memory; "
-                    f"consider narrowing upstream_platforms (platform_instance / env) to "
-                    f"the assets this source references."
-                )
-            else:
-                logger.info(message)
-            self._resolvers_by_platform[platform] = resolvers
-
-    @staticmethod
-    def _strip_platform_instance(name: str, platform_instance: Optional[str]) -> str:
-        # A dataset URN name fuses any platform_instance as a leading ``<instance>.``
-        # prefix; resolve_table re-prepends the resolver's instance, so strip it here
-        # (matched case-insensitively, since that prefix's own casing may differ) to avoid
-        # a doubled prefix.
-        if platform_instance and name.lower().startswith(
-            f"{platform_instance.lower()}."
-        ):
-            return name[len(platform_instance) + 1 :]
-        return name
-
-    def _resolve_dataset(self, urn: str) -> _Resolution:
-        """Resolve `urn` to the casing DataHub already stores, via SchemaResolver.
-
-        Delegates matching to ``SchemaResolver.resolve_table``, which tries three casing
-        candidates for the reference and returns the first that exists in DataHub:
-
-        1. **original** — the reference's name exactly as given.
-        2. **lowercased** — name *and* platform_instance lowercased.
-        3. **mixed** — name lowercased but the platform_instance's casing kept.
-
-        (2) and (3) differ only when a platform_instance has non-lowercase casing. Example,
-        instance ``ProdWarehouse`` stored as ``ProdWarehouse.db.schema.table``, reference
-        ``ProdWarehouse.DB.SCHEMA.TABLE``: (1) misses (table cased wrong), (2) misses
-        (instance lowercased to ``prodwarehouse``), (3) matches.
-
-        A hit under the reference's own casing is EXACT; a hit under a different candidate
-        is NORMALIZED; no hit is UNRESOLVED — the latter includes casings none of the three
-        candidates reach (e.g. an UPPER/Pascal/Mixed-cased *table* in the warehouse), which
-        are the tracked ``normalizedUrn`` SchemaResolver follow-up. The resolved entity's
-        schema is returned too, for column-casing correction.
-        """
-        try:
-            dataset_urn = DatasetUrn.from_string(urn)
-            platform = DataPlatformUrn.from_string(dataset_urn.platform).platform_name
-        except Exception:
-            return _Resolution(urn, None, None)
-        # Track referenced platforms so _warn_unmatched_platforms can flag configured
-        # platforms that no reference used (usually a case/spelling typo in the config).
-        self._seen_reference_platforms.add(platform)
-        resolvers = self._resolvers_by_platform.get(platform)
-        if not resolvers:
-            return _Resolution(urn, None, None)
-
-        name = dataset_urn.name
-        # A platform can have several configured resolvers (one per platform_instance /
-        # env); we iterate and take the first schema match. We deliberately don't index by
-        # platform_instance because it isn't recoverable from the URN — it's fused into the
-        # name, and separating it would require fetching the dataPlatformInstance aspect
-        # (overkill). env *is* recoverable from the URN, so we could additionally
-        # disambiguate on it (e.g. avoid healing a PROD ref to a DEV entity), but that
-        # collision is unlikely in practice, so it's left as a possible follow-up.
-        for resolver in resolvers:
-            table = self._strip_platform_instance(name, resolver.platform_instance)
-            try:
-                # We pass the whole name as `table` and rely on get_urn_for_table
-                # concatenating the parts back into the name. This leans on SchemaResolver
-                # internals and is a bit fragile (get_urn_for_table carries a TODO about
-                # 2/3-layer hierarchy). A read-only resolve_dataset_urn(urn) helper on
-                # SchemaResolver is a tracked follow-up (additive, separate from the
-                # normalizedUrn work).
-                resolved_urn, schema = resolver.resolve_table_parts(
-                    database=None, db_schema=None, table=table
-                )
-            except Exception:
-                continue
-            # resolve_table returns a best-effort URN even on a miss; a non-None schema is
-            # the signal that an existing entity actually matched.
-            if schema is not None:
-                match_type = _EXACT if resolved_urn == urn else _NORMALIZED
-                return _Resolution(resolved_urn, schema, match_type)
-        # On a configured platform but no existing entity matched under a casing
-        # resolve_table covers: leave the URN unchanged but flag it UNRESOLVED so
-        # potentially broken lineage is visible rather than indistinguishable from clean.
-        self.report.unresolved_refs_sample.append(urn)
-        return _Resolution(urn, None, _UNRESOLVED)
-
     # --- aspect rewriters -------------------------------------------------------
     #
     # Each returns True iff it mutated the aspect (rewrote a reference or stamped a
     # matchType), so process() can skip the raw-MCP re-serialization when nothing in the
     # aspect was in scope.
 
-    def _tally_table_ref(self, res: _Resolution) -> bool:
+    def _tally_table_ref(self, res: Resolution) -> bool:
         """Record report counters for a table-level reference; return True iff it was
         normalized (so the caller can rewrite the URN). Shared by the three table-level
         paths; the column-level path (_resolve_field_urn) counts separately."""
-        if res.match_type == _NORMALIZED:
+        if res.match_type == NORMALIZED:
             self.report.num_dataset_urns_normalized += 1
             return True
-        if res.match_type == _UNRESOLVED:
+        if res.match_type == UNRESOLVED:
             self.report.num_refs_unresolved += 1
+            self.report.unresolved_refs_sample.append(res.urn)
         else:
             self.report.num_refs_unchanged += 1
         return False
@@ -558,10 +315,9 @@ class AutoResolveLineageUrnsProcessor(
             dataset = getattr(upstream, "dataset", None)
             if not _is_dataset_urn(dataset):
                 continue
-            res = self._resolve_dataset(dataset)
-            # Stamp the verdict (EXACT / NORMALIZED / UNRESOLVED) for any reference on
-            # a configured platform; out-of-scope refs get res.match_type=None and are
-            # left untouched.
+            res = self._strategy.resolve(dataset)
+            # Stamp the verdict (EXACT / NORMALIZED / UNRESOLVED) for any reference in
+            # scope; out-of-scope refs get res.match_type=None and are left untouched.
             if res.match_type is not None:
                 upstream.matchType = res.match_type
                 changed = True
@@ -601,12 +357,12 @@ class AutoResolveLineageUrnsProcessor(
         # field couldn't be matched) > EXACT (all verified). Absent only when every
         # field was out of scope.
         aggregate: Optional[str] = None
-        if _NORMALIZED in match_types:
-            aggregate = _NORMALIZED
-        elif _UNRESOLVED in match_types:
-            aggregate = _UNRESOLVED
-        elif _EXACT in match_types:
-            aggregate = _EXACT
+        if NORMALIZED in match_types:
+            aggregate = NORMALIZED
+        elif UNRESOLVED in match_types:
+            aggregate = UNRESOLVED
+        elif EXACT in match_types:
+            aggregate = EXACT
         if aggregate is not None:
             fine_grained.matchType = aggregate
             changed = True
@@ -620,14 +376,15 @@ class AutoResolveLineageUrnsProcessor(
             return field_urn, None
 
         # Column-level: we need the parent's schema to correct the column casing.
-        res = self._resolve_dataset(parent)
+        res = self._strategy.resolve(parent, need_schema=True)
         new_field_path = field_path
         if res.schema:
             new_field_path = self._match_columns_to_schema(res.schema, [field_path])[0]
 
         if res.urn == parent and new_field_path == field_path:
-            if res.match_type == _UNRESOLVED:
+            if res.match_type == UNRESOLVED:
                 self.report.num_refs_unresolved += 1
+                self.report.unresolved_refs_sample.append(res.urn)
             else:
                 self.report.num_refs_unchanged += 1
             return field_urn, res.match_type
@@ -638,7 +395,7 @@ class AutoResolveLineageUrnsProcessor(
         # even when the parent dataset matched exactly, so report NORMALIZED in that
         # case rather than the parent's (EXACT) match type.
         self.report.num_column_urns_normalized += 1
-        match_type = _NORMALIZED if new_field_path != field_path else res.match_type
+        match_type = NORMALIZED if new_field_path != field_path else res.match_type
         return make_schema_field_urn(res.urn, new_field_path), match_type
 
     def _normalize_dashboard_info(self, aspect: DashboardInfoClass) -> bool:
@@ -689,7 +446,7 @@ class AutoResolveLineageUrnsProcessor(
             if not _is_dataset_urn(dataset):
                 healed.append(dataset)
                 continue
-            res = self._resolve_dataset(dataset)
+            res = self._strategy.resolve(dataset)
             # A plain URN list / Edge has no matchType field to stamp, but the counters
             # must still distinguish UNRESOLVED (broken) from a clean ref so a
             # dashboard/datajob pointing at broken lineage isn't invisible in the report.
@@ -704,7 +461,7 @@ class AutoResolveLineageUrnsProcessor(
             destination = getattr(edge, "destinationUrn", None)
             if not _is_dataset_urn(destination):
                 continue
-            res = self._resolve_dataset(destination)
+            res = self._strategy.resolve(destination)
             if self._tally_table_ref(res):
                 edge.destinationUrn = res.urn
                 changed = True
